requirements.txt
-i https://pypi.python.org/simple
certifi==2018.4.16
chardet==3.0.4
click==6.7
flask==1.0.2
future==0.16.0
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
python-telegram-bot==10.1.0
requests==2.18.4
urllib3==1.22
werkzeug==0.14.1
